+++ kind-integer/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Version 0.1.1
 
-* Export `Abs`, `GCD`, `LCM`, `Odd`, `Even`, `toPrelude`, `fromPrelude`.
+* Export `Abs`, `GCD`, `LCM`, `Odd`, `Even`, `toPrelude`, `fromPrelude`,
+  `showsPrecTypeLit`.
+
+* Add `Eq`, `Ord`, `Show`, `Read` instances for `Integer`.
 
 * Minor cabal and documentation improvements.